pyproject.toml
[project]
name = "sql-query-optimization-benchmarking"
version = "0.1.0"
description = "SQL query execution behavior analysis and performance benchmarking"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "venv", "env", ".venv"]

